requirements.txt
msrestazure>=0.4.14
azure-mgmt-resource>=1.1.0
azure-keyvault